Refactor imports to include validator across multiple modules for improved consistency
