+++ integrations/slack/events/comparison.py
@@ -6,12 +6,12 @@ import logging
 from typing import TYPE_CHECKING, cast
 
 import libs.global_value as g
-from libs.domain import modify
+from libs.domain import modify, validator
 from libs.domain.datamodels import ComparisonResults
 from libs.domain.score import GameResult
 from libs.functions import lookup, search
 from libs.types import ActionStatus, CommandType, StyleOptions
-from libs.utils import formatter, validator
+from libs.utils import formatter
 from libs.utils.timekit import ExtendedDatetime as ExtDt
 
 if TYPE_CHECKING:
