+++ libs/dispatcher.py
@@ -7,11 +7,11 @@ from typing import TYPE_CHECKING
 
 import libs.global_value as g
 from integrations import factory
-from libs.domain import modify
+from libs.domain import modify, validator
 from libs.domain.score import GameResult
 from libs.functions import lookup, message
 from libs.types import MessageStatus, StyleOptions
-from libs.utils import formatter, validator
+from libs.utils import formatter
 
 if TYPE_CHECKING:
     from integrations.protocols import MessageParserProtocol
