+++ libs/domain/__init__.py
@@ -9,4 +9,5 @@
 - :doc:`libs.domain.placeholder`: プレースホルダ構築
 - :doc:`libs.domain.aggregate`: データ集計
 - :doc:`libs.domain.modify`: DBデータ操作
+- :doc:`libs.domain.validator`: バリデータ
 """
