+++ libs/functions/tools/unification.py
@@ -6,9 +6,9 @@ import configparser
 import logging
 
 import libs.global_value as g
-from libs.domain import modify
+from libs.domain import modify, validator
 from libs.functions import lookup
-from libs.utils import dbutil, textutil, validator
+from libs.utils import dbutil, textutil
 
 
 def main() -> None:
